Update Timelight.js
Added conditional to draw the clock if useLocalTime is true and the blackWhiteGradient falls in the range [120, 130]. The draw methods inside the conditional have the gradient passed with a value of plus or minus 5 so the distinction between dark and light is more prominent. Without this, the screen will briefly be entirely gray with no discernable features.